requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
paddleocr
paddlepaddle-gpu
ultralytics
sentence-transformers
numpy<2.0.0
torch
